Backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.0
bcrypt==4.1.2
google-auth==2.27.0